Add middleware and associations

+++ api/server/controllers/operationController.ts
@@ -1,11 +1,17 @@
-import { Request, Response } from "express";
+import { Response } from "express";
+import { IReqAuth } from "../utils/types";
 import Operations, { IOperation } from "../models/operationModel";
 
-export const getOperations = async (req: Request, res: Response) => {
+export const getOperations = async (req: IReqAuth, res: Response) => {
     try {
 
+        if (!req.user) return res.status(400).json({ msg: 'Invalid authentication.' });
+
         const operations: IOperation[] | null = await Operations.findAll({
-            limit: 10
+            limit: 10,
+            where: {
+                userId: req.user.id
+            }
         });
 
         return res.status(200).json({ operations });
@@ -15,12 +21,15 @@ export const getOperations = async (req: Request, res: Response) => {
     }
 };
 
-export const createOperation = async (req: Request, res: Response) => {
+export const createOperation = async (req: IReqAuth, res: Response) => {
+
+    if (!req.user) return res.status(400).json({ msg: 'Invalid authentication.' });
+
     try {
 
         const { description, amount, type } = req.body;
 
-        const newOperation = await Operations.create({ description, amount, type });
+        const newOperation = await Operations.create({ description, amount, type, userId: req.user.id });
 
         return res.status(201).json({ msg: 'Operation created successfully.', newOperation });
 
@@ -29,7 +38,10 @@ export const createOperation = async (req: Request, res: Response) => {
     }
 };
 
-export const updateOperation = async (req: Request, res: Response) => {
+export const updateOperation = async (req: IReqAuth, res: Response) => {
+
+    if (!req.user) return res.status(400).json({ msg: 'Invalid authentication.' });
+
     try {
 
         const { description, amount } = req.body;
@@ -40,7 +52,7 @@ export const updateOperation = async (req: Request, res: Response) => {
 
         await Operations.update(
             { description, amount }, {
-            where: { id: req.params.id }
+            where: { id: req.params.id, userId: req.user.id }
         });
 
         return res.status(200).json({ msg: 'Operation updated successfully.' });
@@ -50,12 +62,16 @@ export const updateOperation = async (req: Request, res: Response) => {
     }
 };
 
-export const deleteOperation = async (req: Request, res: Response) => {
+export const deleteOperation = async (req: IReqAuth, res: Response) => {
+
+    if (!req.user) return res.status(400).json({ msg: 'Invalid authentication.' });
+
     try {
 
         await Operations.destroy({
             where: {
-                id: req.params.id
+                id: req.params.id,
+                userId: req.user.id
             }
         });
